model/requirements.txt
requests
numpy
scikit-learn
google-cloud-storage
nltk